% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Double.R
\name{Double}
\alias{Double}
\title{Double}
\usage{
Double(number)
}
\arguments{
\item{number}{A numerical vector}
}
\value{
A vector containing the doubles of the vector
}
\description{
compute double of vector
}
\details{
the double is calculated by multiplying each value of vector by 2
}
\examples{
Double(runinf(3))
}
